+++ include/drivers/apic/apic.h
@@ -58,7 +58,7 @@ void init_lapic(void);
 size_t get_cpunum(void);
 void write_lapic(size_t index, uint32_t value);
 
-void start_ap(uint8_t apicid, uintptr_t addr);
+void start_ap(size_t apicid, uintptr_t addr);
 
 } // namespace local_apic
